✨ feat(postgresql): add MockRow, MockRows, and MockResult types
- implement MockRow for mocking QueryRow responses
- implement MockRows for mocking pgx.Rows with added Values() method
- implement MockResult for mocking CommandTag results
- enhance testing capabilities for GoodsPGRepository

🔧 chore(repository_test): refactor tests to use new mocks

- replace MockRow with MockRows in TestGetProductByID
- update TestGetAllProducts to use MockRows
- replace MockRow with new MockRow in CreateProduct tests
- streamline setup of mock client in tests using newTestRepo() function
