+++ internal/repository/postgresql/mock_client.go
@@ -8,6 +8,18 @@ import (
 	"github.com/stretchr/testify/mock"
 )
 
+// Тест 1 --------------------------------------------------------------------------------
+
+// MockRow для мокирования ответа QueryRow
+type MockRow struct {
+	mock.Mock
+}
+
+func (m *MockRow) Scan(dest ...interface{}) error {
+	args := m.Called(dest...)
+	return args.Error(0)
+}
+
 // MockClient — мок для тестирования GoodsPGRepository
 type MockClient struct {
 	mock.Mock
@@ -50,4 +62,60 @@ func (m *MockClient) Close() {
 // Добавляем Ping, если используется
 func (m *MockClient) Ping(ctx context.Context) error {
 	return m.Called(ctx).Error(0)
+}
+
+// Тест 2 --------------------------------------------------------------------------------
+
+// MockRows для мокирования pgx.Rows
+type MockRows struct {
+	mock.Mock
+}
+
+func (m *MockRows) Next() bool {
+	args := m.Called()
+	return args.Bool(0)
+}
+
+func (m *MockRows) Scan(dest ...interface{}) error {
+	args := m.Called(dest...)
+	return args.Error(0)
+}
+
+func (m *MockRows) Err() error {
+	args := m.Called()
+	return args.Error(0)
+}
+
+func (m *MockRows) CommandTag() pgconn.CommandTag {
+	return pgconn.CommandTag{}
+}
+
+func (m *MockRows) Conn() *pgx.Conn {
+	return nil
+}
+
+func (m *MockRows) Close() {}
+func (m *MockRows) FieldDescriptions() []pgconn.FieldDescription {
+	return nil
+}
+func (m *MockRows) RawValues() [][]byte {
+	return nil
+}
+
+// Исправление: добавляем метод Values()
+func (m *MockRows) Values() ([]interface{}, error) {
+	args := m.Called()
+	return args.Get(0).([]interface{}), args.Error(1)
+}
+
+// Тест 4 --------------------------------------------------------------------------------
+
+// MockResult для мокирования CommandTag (результата Exec)
+type MockResult struct {
+	mock.Mock
+}
+
+func (m *MockResult) RowsAffected() int64 {
+	args := m.Called()
+	return args.Get(0).(int64)
 }
